chore: refactor utils file to /front/lib
