+++ front/app/api/proposal/create/service.ts
@@ -1,7 +1,11 @@
 import * as anchor from "@coral-xyz/anchor";
-import { PublicKey } from "@solana/web3.js";
-import { getProgram, getAdminKeypair, createAnchorWallet, RPC_ENDPOINT } from "../../shared/utils";
-import { Connection } from "@solana/web3.js";
+import { Connection, PublicKey } from "@solana/web3.js";
+import {
+  createAnchorWallet,
+  getAdminKeypair,
+  getProgram,
+  RPC_ENDPOINT,
+} from "../../../../lib/utils";
 
 export interface ProposalCreateParams {
   tokenName: string;
@@ -15,8 +19,11 @@ export interface ProposalCreateParams {
 }
 
 export async function createProposal(params: ProposalCreateParams) {
-  console.log("🔍 Initialisation de la création de proposition avec params:", params);
-  
+  console.log(
+    "🔍 Initialisation de la création de proposition avec params:",
+    params
+  );
+
   const connection = new Connection(RPC_ENDPOINT);
   const adminKeypair = getAdminKeypair();
   const wallet = createAnchorWallet(adminKeypair);
@@ -26,7 +33,10 @@ export async function createProposal(params: ProposalCreateParams) {
     throw new Error("Programme non initialisé");
   }
 
-  console.log("✅ Programme initialisé, autorité:", adminKeypair.publicKey.toString());
+  console.log(
+    "✅ Programme initialisé, autorité:",
+    adminKeypair.publicKey.toString()
+  );
 
   // Récupérer l'ID d'époque actif si aucun n'est fourni
   let epochId = params.epochId;
@@ -37,7 +47,7 @@ export async function createProposal(params: ProposalCreateParams) {
     // Récupérer toutes les époques disponibles
     const allEpochs = await (program.account as any).epochManagement.all();
     console.log("✅ Nombre d'époques trouvées:", allEpochs.length);
-    
+
     if (allEpochs.length === 0) {
       throw new Error("Aucune époque trouvée dans le programme");
     }
@@ -45,25 +55,25 @@ export async function createProposal(params: ProposalCreateParams) {
     // Si epochId est spécifié, chercher cette époque spécifique
     if (epochId) {
       console.log("🔍 Recherche de l'époque spécifique:", epochId);
-      const specificEpoch = allEpochs.find((epoch: any) => 
-        epoch.account.epochId.toString() === epochId
+      const specificEpoch = allEpochs.find(
+        (epoch: any) => epoch.account.epochId.toString() === epochId
       );
-      
+
       if (!specificEpoch) {
         throw new Error(`L'époque avec l'ID ${epochId} n'existe pas`);
       }
-      
+
       epochPDA = specificEpoch.publicKey;
       epochAccount = specificEpoch.account;
     } else {
       // Sinon, prendre la première époque active ou la plus récente
       console.log("🔍 Recherche d'une époque active par défaut...");
-      
+
       // D'abord, essayer de trouver une époque active
-      const activeEpoch = allEpochs.find((epoch: any) => 
-        epoch.account.status.active !== undefined
+      const activeEpoch = allEpochs.find(
+        (epoch: any) => epoch.account.status.active !== undefined
       );
-      
+
       if (activeEpoch) {
         epochPDA = activeEpoch.publicKey;
         epochAccount = activeEpoch.account;
@@ -73,27 +83,31 @@ export async function createProposal(params: ProposalCreateParams) {
         // Sinon, prendre la dernière époque créée
         console.log("⚠️ Aucune époque active, sélection de la dernière époque");
         // Trier par ID d'époque (supposant que des IDs plus élevés sont plus récents)
-        allEpochs.sort((a: any, b: any) => 
+        allEpochs.sort((a: any, b: any) =>
           b.account.epochId.cmp(a.account.epochId)
         );
-        
+
         epochPDA = allEpochs[0].publicKey;
         epochAccount = allEpochs[0].account;
         epochId = epochAccount.epochId.toString();
         console.log("✅ Dernière époque sélectionnée:", epochId);
       }
     }
-    
+
     console.log("✅ Époque sélectionnée:", {
       id: epochId,
       pubkey: epochPDA.toString(),
-      status: Object.keys(epochAccount.status)[0]
+      status: Object.keys(epochAccount.status)[0],
     });
   } catch (error) {
     console.error("❌ Erreur lors de la récupération des époques:", error);
-    throw new Error(`Impossible de trouver une époque valide: ${error instanceof Error ? error.message : String(error)}`);
+    throw new Error(
+      `Impossible de trouver une époque valide: ${
+        error instanceof Error ? error.message : String(error)
+      }`
+    );
   }
-  
+
   // Vérifier que le trésor est initialisé
   const [treasuryPDA] = PublicKey.findProgramAddressSync(
     [Buffer.from("treasury")],
@@ -106,7 +120,7 @@ export async function createProposal(params: ProposalCreateParams) {
       Buffer.from("proposal"),
       adminKeypair.publicKey.toBuffer(),
       new anchor.BN(epochAccount.epochId).toArrayLike(Buffer, "le", 8),
-      Buffer.from(params.tokenName)
+      Buffer.from(params.tokenName),
     ],
     program.programId
   );
@@ -136,7 +150,9 @@ export async function createProposal(params: ProposalCreateParams) {
   console.log("✅ Transaction envoyée:", tx);
 
   // Récupérer les détails de la proposition créée
-  const proposalAccount = await (program.account as any).tokenProposal.fetch(proposalPDA);
+  const proposalAccount = await (program.account as any).tokenProposal.fetch(
+    proposalPDA
+  );
 
   return {
     transactionId: tx,
@@ -152,6 +168,6 @@ export async function createProposal(params: ProposalCreateParams) {
       status: Object.keys(proposalAccount.status)[0],
       creationTimestamp: proposalAccount.creationTimestamp.toString(),
       lockupPeriod: proposalAccount.lockupPeriod.toString(),
-    }
+    },
   };
-} 
+}
