+++ front/app/api/treasury/balance/route.ts
@@ -1,9 +1,16 @@
 // front/app/api/treasury/balance/route.ts
+import { Connection, PublicKey } from "@solana/web3.js";
+import { randomUUID } from "crypto";
 import { NextRequest } from "next/server";
-import { randomUUID } from 'crypto';
-import { verifyAuthToken, createSuccessResponse, createErrorResponse } from "../../shared/utils";
-import { getProgram, getAdminKeypair, createAnchorWallet, RPC_ENDPOINT } from "../../shared/utils";
-import { PublicKey, Connection } from "@solana/web3.js";
+import {
+  createAnchorWallet,
+  createErrorResponse,
+  createSuccessResponse,
+  getAdminKeypair,
+  getProgram,
+  RPC_ENDPOINT,
+  verifyAuthToken,
+} from "../../../../lib/utils";
 
 export async function GET(request: NextRequest): Promise<Response> {
   const requestId = randomUUID();
@@ -12,10 +19,14 @@ export async function GET(request: NextRequest): Promise<Response> {
   // Vérification du token d'authentification
   if (!verifyAuthToken(request)) {
     console.error(`[${requestId}] ❌ Authentification échouée`);
-    return createErrorResponse(requestId, {
-      message: "Non autorisé",
-      name: "AuthenticationError"
-    }, 401);
+    return createErrorResponse(
+      requestId,
+      {
+        message: "Non autorisé",
+        name: "AuthenticationError",
+      },
+      401
+    );
   }
 
   try {
@@ -35,40 +46,57 @@ export async function GET(request: NextRequest): Promise<Response> {
     );
 
     // Récupérer le compte treasury
-    const treasuryAccount = await (program.account as any).treasury.fetchNullable(treasuryPDA);
-    
+    const treasuryAccount = await (
+      program.account as any
+    ).treasury.fetchNullable(treasuryPDA);
+
     if (!treasuryAccount) {
-      return createErrorResponse(requestId, {
-        message: "La treasury n'est pas initialisée.",
-      }, 404);
+      return createErrorResponse(
+        requestId,
+        {
+          message: "La treasury n'est pas initialisée.",
+        },
+        404
+      );
     }
 
     // Récupérer le solde direct du compte PDA sur la blockchain et convertir en SOL
-    const treasuryPdaBalanceSOL = await connection.getBalance(treasuryPDA) / 1_000_000_000;
+    const treasuryPdaBalanceSOL =
+      (await connection.getBalance(treasuryPDA)) / 1_000_000_000;
 
     // Accéder aux balances de sous-comptes avec la même notation que dans les tests et convertir en SOL
-    const marketingBalanceSOL = (treasuryAccount.marketing.solBalance ? 
-      treasuryAccount.marketing.solBalance.toNumber() : 
-      treasuryAccount.marketing.sol_balance) / 1_000_000_000;
-      
-    const teamBalanceSOL = (treasuryAccount.team.solBalance ? 
-      treasuryAccount.team.solBalance.toNumber() : 
-      treasuryAccount.team.sol_balance) / 1_000_000_000;
-      
-    const operationsBalanceSOL = (treasuryAccount.operations.solBalance ? 
-      treasuryAccount.operations.solBalance.toNumber() : 
-      treasuryAccount.operations.sol_balance) / 1_000_000_000;
-      
-    const investmentsBalanceSOL = (treasuryAccount.investments.solBalance ? 
-      treasuryAccount.investments.solBalance.toNumber() : 
-      treasuryAccount.investments.sol_balance) / 1_000_000_000;
-      
-    const crankBalanceSOL = (treasuryAccount.crank.solBalance ? 
-      treasuryAccount.crank.solBalance.toNumber() : 
-      treasuryAccount.crank.sol_balance) / 1_000_000_000;
+    const marketingBalanceSOL =
+      (treasuryAccount.marketing.solBalance
+        ? treasuryAccount.marketing.solBalance.toNumber()
+        : treasuryAccount.marketing.sol_balance) / 1_000_000_000;
+
+    const teamBalanceSOL =
+      (treasuryAccount.team.solBalance
+        ? treasuryAccount.team.solBalance.toNumber()
+        : treasuryAccount.team.sol_balance) / 1_000_000_000;
+
+    const operationsBalanceSOL =
+      (treasuryAccount.operations.solBalance
+        ? treasuryAccount.operations.solBalance.toNumber()
+        : treasuryAccount.operations.sol_balance) / 1_000_000_000;
+
+    const investmentsBalanceSOL =
+      (treasuryAccount.investments.solBalance
+        ? treasuryAccount.investments.solBalance.toNumber()
+        : treasuryAccount.investments.sol_balance) / 1_000_000_000;
+
+    const crankBalanceSOL =
+      (treasuryAccount.crank.solBalance
+        ? treasuryAccount.crank.solBalance.toNumber()
+        : treasuryAccount.crank.sol_balance) / 1_000_000_000;
 
     // Calculer le total des sous-comptes en SOL
-    const totalSubAccountsSOL = marketingBalanceSOL + teamBalanceSOL + operationsBalanceSOL + investmentsBalanceSOL + crankBalanceSOL;
+    const totalSubAccountsSOL =
+      marketingBalanceSOL +
+      teamBalanceSOL +
+      operationsBalanceSOL +
+      investmentsBalanceSOL +
+      crankBalanceSOL;
 
     // Log des balances en SOL dans le format des tests
     console.log("\nVérification Soldes (Treasury) en SOL:");
@@ -82,7 +110,7 @@ export async function GET(request: NextRequest): Promise<Response> {
     console.log(`  - Investments: ${investmentsBalanceSOL} SOL`);
     console.log(`  - Crank: ${crankBalanceSOL} SOL`);
     console.log(`  - Total sous-comptes: ${totalSubAccountsSOL} SOL`);
-    
+
     return createSuccessResponse(requestId, {
       // Message sur le statut
       message: "Treasury existante et accessible",
@@ -94,15 +122,18 @@ export async function GET(request: NextRequest): Promise<Response> {
         operations: operationsBalanceSOL,
         investments: investmentsBalanceSOL,
         crank: crankBalanceSOL,
-        total: totalSubAccountsSOL
+        total: totalSubAccountsSOL,
       },
       // Informations sur le compte
       address: treasuryPDA.toString(),
       authority: treasuryAccount.authority.toString(),
-      RPC_ENDPOINT: RPC_ENDPOINT
+      RPC_ENDPOINT: RPC_ENDPOINT,
     });
   } catch (error) {
-    console.error(`[${requestId}] ❌ Erreur lors de la récupération de la balance:`, error);
+    console.error(
+      `[${requestId}] ❌ Erreur lors de la récupération de la balance:`,
+      error
+    );
     return createErrorResponse(requestId, error);
   }
 }
